Mark the unhash/hash functions static
Marking them static tells the compiler that those
functions will only be called from inside this file.
This means that it can make assumptions based on how
it is called - in this case, it recognizes that the arguments
to unhash are the same and bakes them in.  Since hash is inlined,
this allows it to unroll the loop in hash.  Finally, because of the
unrolling and argument baking, it can strip out some unnecessary
arithmetic and just subtract 7*37^(length) from the target.

Time is now .29 seconds.